Cleaning up main.h header file
