+++ embedded/common/app/cmdhandler_t.c
@@ -24,6 +24,7 @@
 #include "common.h"
 #include "hw_setup.h"
 #include <string.h>
+#include "debugprint.h"
 
 
 /*-------------------------------------------------------------------------------------------------*\
@@ -39,6 +40,8 @@ extern void CmdParse_User( int8_t *pBuffer, uint16_t size );
 /*-------------------------------------------------------------------------------------------------*\
  |    P R I V A T E   C O N S T A N T S   &   M A C R O S
 \*-------------------------------------------------------------------------------------------------*/
+/* Defines for command handler */
+#define COMMAND_LINE_SIZE                       32
 
 /*-------------------------------------------------------------------------------------------------*\
  |    P R I V A T E   T Y P E   D E F I N I T I O N S
