+++ embedded/projects/osp-lpc54102/sources/app/console_util.c
@@ -36,6 +36,16 @@
 /*-------------------------------------------------------------------------------------------------*\
  |    P R I V A T E   T Y P E   D E F I N I T I O N S
 \*-------------------------------------------------------------------------------------------------*/
+/* Serial command parser tokens */
+enum ParserTokensTag {
+    TOKEN_NULL = 0,
+    TOKEN_1 = 'l',
+    TOKEN_2 = 'o',
+    TOKEN_3 = 'g',
+    TOKEN_4 = '=',
+    TOKEN_STATS = '\r',
+    TOKEN_PARAM = 0xAA
+};
 
 /*-------------------------------------------------------------------------------------------------*\
  |    S T A T I C   V A R I A B L E S   D E F I N I T I O N S
