+++ embedded/projects/osp-lpc54102/sources/boardsupport/hw_setup.c
@@ -20,6 +20,7 @@
 \*-------------------------------------------------------------------------------------------------*/
 #include "common.h"
 #include "hw_setup.h"
+#include "debugprint.h"
 
 /*-------------------------------------------------------------------------------------------------*\
  |    E X T E R N A L   V A R I A B L E S   &   F U N C T I O N S
