+++ embedded/projects/osp-lpc54102/sources/boardsupport/sensacq_i2c.h
@@ -32,6 +32,33 @@
 #ifndef __SENSACQ_I2C_H
 #define __SENSACQ_I2C_H
 #include "osp-types.h"
+
+#define INTERRUPT_BASED_SAMPLING    /* Sensor sampling is interrupt driver */
+#undef  TRIGGERED_MAG_SAMPLING      /* Magnetometer sampling is software triggered */
+/* Sensor acquisition related definitions */
+
+#if !defined INTERRUPT_BASED_SAMPLING
+# define SENSOR_SAMPLE_PERIOD           (20)   //time in ms
+# define MAG_DECIMATE_FACTOR            (1)
+# define ACCEL_SAMPLE_DECIMATE          (1)
+# define GYRO_SAMPLE_DECIMATE           (1)
+#else
+# define MAG_DECIMATE_FACTOR            (1)
+# define ACCEL_SAMPLE_DECIMATE          (1)
+# define GYRO_SAMPLE_DECIMATE           (1)
+#endif
+
+#define	PRESSURE_SAMPLE_PERIOD		(40)
+
+#ifdef TRIGGERED_MAG_SAMPLING
+# define MAG_TRIGGER_RATE_DECIMATE      (1) //1/2 of Accel ODR
+#endif
+
+#define ALGORITHM_TASK                  ALGORITHM_TASK_ID
+
+#define TIMER_REF_SENSOR_READ               (0x55B0)
+#define TIMER_REF_PRESSURE_READ             (0x55ca)
+
 osp_bool_t dev_i2c_init(void);
 
 void dev_i2c_delay(unsigned int msec);
